# Improve the results from the faster-greedy-dag extractor