@oflatt 2024-01-06
This is an interesting experiment- not sure if we should merge it since we don't have a reason why it's better
Maybe you could merge it as faster-greedy-dag-alt? Or just keep this a draft.