@mwillsey 2024-01-12
Yeah I think we close this for now, unless something changes and makes it better. Fewer more distinct extractors seems like the way to go for now. Happy to re-open if other feel differently.